feat: 'SFDX: Turn On Apex Debug Log for Replay Debugger' works without CLI
