+++ packages/salesforcedx-vscode-core/src/commands/index.ts
@@ -35,7 +35,8 @@ export { projectDeployStart } from './projectDeployStart';
 export { projectGenerateWithManifest, sfProjectGenerate } from './projectGenerate';
 export { projectRetrieveStart } from './projectRetrieveStart';
 export { viewAllChanges, viewLocalChanges, viewRemoteChanges } from './source/viewChanges';
-export { startApexDebugLogging } from './startApexDebugLogging';
+export { startApexDebugLogging } from './startApexDebugLoggingOld';
+export { turnOnLogging } from './startApexDebugLogging';
 export { turnOffLogging } from './stopApexDebugLogging';
 export { taskStop } from './taskStop';
 export {
