+++ packages/salesforcedx-vscode-core/src/commands/startApexDebugLogging.ts
@@ -4,98 +4,80 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { CommandOutput, CommandBuilder, Command, SfCommandBuilder } from '@salesforce/salesforcedx-utils';
-import {
-  CliCommandExecutor,
-  CommandExecution,
-  CompositeCliCommandExecutor,
-  ContinueResponse,
-  EmptyParametersGatherer,
-  workspaceUtils
-} from '@salesforce/salesforcedx-utils-vscode';
-import * as vscode from 'vscode';
-import { channelService } from '../channels';
+
+import { CliCommandExecutor, CommandOutput } from '@salesforce/salesforcedx-utils';
 import { APEX_CODE_DEBUG_LEVEL, VISUALFORCE_DEBUG_LEVEL } from '../constants';
-import { workspaceContextUtils } from '../context';
+import { WorkspaceContext, workspaceContextUtils } from '../context';
 import { nls } from '../messages';
 import { telemetryService } from '../telemetry';
 import { OrgAuthInfo } from '../util';
+import { handleStartCommand, handleFinishCommand } from '../utils/channelUtils';
 import { developerLogTraceFlag } from '.';
-import { SfCommandlet, SfCommandletExecutor, SfWorkspaceChecker } from './util';
+import { QueryUser } from './startApexDebugLoggingOld';
+import { workspaceUtils } from '@salesforce/salesforcedx-utils-vscode';
 
-class StartApexDebugLoggingExecutor extends SfCommandletExecutor<{}> {
-  private cancellationTokenSource = new vscode.CancellationTokenSource();
-  private cancellationToken = this.cancellationTokenSource.token;
+const command = 'start_apex_debug_logging';
 
-  public build(): Command {
-    return new CommandBuilder(nls.localize('start_apex_debug_logging')).withLogName('start_apex_debug_logging').build();
-  }
+// const generateTraceFlagId = (): string => {
+//   const timestamp = Date.now().toString();
+//   const random = Math.floor(Math.random() * 1000)
+//     .toString()
+//     .padStart(3, '0');
+//   return `7tf${timestamp}${random}`;
+// };
 
-  public attachSubExecution(execution: CommandExecution) {
-    channelService.streamCommandOutput(execution);
-  }
+export const turnOnLogging = async (): Promise<void> => {
+  handleStartCommand(command);
 
-  public async execute(response: ContinueResponse<{}>): Promise<void> {
-    const startTime = process.hrtime();
-    const executionWrapper = new CompositeCliCommandExecutor(this.build()).execute(this.cancellationToken);
-    this.attachExecution(executionWrapper, this.cancellationTokenSource, this.cancellationToken);
+  try {
+    const connection = await WorkspaceContext.getInstance().getConnection();
 
-    executionWrapper.processExitSubject.subscribe(() => {
-      this.logMetric(executionWrapper.command.logName, startTime);
-    });
-
-    try {
-      // query traceflag
-      const userId = await getUserId(workspaceUtils.getRootWorkspacePath());
-
-      let resultJson = await this.subExecute(new QueryTraceFlag().build(userId));
-      if (resultJson && resultJson.result && resultJson.result.totalSize >= 1) {
-        const traceflag = resultJson.result.records[0];
-        developerLogTraceFlag.setTraceFlagDebugLevelInfo(
-          traceflag.Id,
-          traceflag.StartDate,
-          traceflag.ExpirationDate,
-          traceflag.DebugLevelId
-        );
-        if (!developerLogTraceFlag.isValidDebugLevelId()) {
-          throw new Error(nls.localize('invalid_debug_level_id_error'));
-        }
-        await this.subExecute(new UpdateDebugLevelsExecutor().build());
-
-        if (!developerLogTraceFlag.isValidDateLength()) {
-          developerLogTraceFlag.validateDates();
-          await this.subExecute(new UpdateTraceFlagsExecutor().build());
-        }
-      } else {
-        resultJson = await this.subExecute(new CreateDebugLevel().build());
-        if (resultJson) {
-          const debugLevelId = resultJson.result.id;
-          developerLogTraceFlag.setDebugLevelId(debugLevelId);
-
-          developerLogTraceFlag.validateDates();
-          resultJson = await this.subExecute(new CreateTraceFlag(userId).build());
-          developerLogTraceFlag.setTraceFlagId(resultJson.result.id);
-        }
-      }
-      developerLogTraceFlag.turnOnLogging();
-      executionWrapper.successfulExit();
-    } catch (e) {
-      executionWrapper.failureExit(e);
+    // Create a new DebugLevel first
+    const debugLevel = {
+      DeveloperName: `ReplayDebuggerLevels${Date.now()}`,
+      MasterLabel: `ReplayDebuggerLevels${Date.now()}`,
+      ApexCode: APEX_CODE_DEBUG_LEVEL,
+      Visualforce: VISUALFORCE_DEBUG_LEVEL
+    };
+    const debugLevelResult = await connection.tooling.create('DebugLevel', debugLevel);
+    if (!debugLevelResult.success) {
+      throw new Error('Failed to create debug level');
     }
-  }
 
-  private async subExecute(command: Command) {
-    if (!this.cancellationToken.isCancellationRequested) {
-      const execution = new CliCommandExecutor(command, {
-        cwd: workspaceUtils.getRootWorkspacePath()
-      }).execute(this.cancellationToken);
-      this.attachSubExecution(execution);
-      const resultPromise = new CommandOutput().getCmdResult(execution);
-      const result = await resultPromise;
-      return JSON.parse(result);
+    // const traceFlag = {
+    //   // TracedEntityId: generateTraceFlagId(),
+    //   TracedEntityId: await getUserId(workspaceUtils.getRootWorkspacePath()),
+    //   LogType: 'DEVELOPER_LOG',
+    //   StartDate: developerLogTraceFlag.getStartDate().toISOString(),
+    //   ExpirationDate: developerLogTraceFlag.getExpirationDate().toISOString(),
+    //   DebugLevelId: debugLevelResult.id
+    // };
+    const expirationDate = new Date(Date.now() + 24 * 60 * 60 * 1000);
+    const traceFlag = {
+      TracedEntityId: await getUserId(workspaceUtils.getRootWorkspacePath()),
+      LogType: 'DEVELOPER_LOG',
+      StartDate: '', // Empty string, as in CLI
+      ExpirationDate: expirationDate.toUTCString(), // RFC1123 format
+      DebugLevelId: debugLevelResult.id
+    };
+    console.log('TraceFlag object:', JSON.stringify(traceFlag, null, 2));
+
+    const traceFlagResult = await connection.tooling.create('TraceFlag', traceFlag);
+    if (!traceFlagResult.success) {
+      throw new Error('Failed to create trace flag');
     }
+
+    developerLogTraceFlag.setTraceFlagId(traceFlagResult.id);
+    developerLogTraceFlag.setDebugLevelId(debugLevelResult.id);
+    developerLogTraceFlag.turnOnLogging();
+
+    await handleFinishCommand(command, true);
+  } catch (error) {
+    console.error(error);
+    await handleFinishCommand(command, false, 'Cannot create trace flag.');
+    throw new Error('Cannot create trace flag.');
   }
-}
+};
 
 export const getUserId = async (projectPath: string): Promise<string> => {
   const targetOrgOrAlias = await workspaceContextUtils.getTargetOrgOrAlias();
@@ -124,119 +106,4 @@ export const getUserId = async (projectPath: string): Promise<string> => {
   } catch {
     return Promise.reject(result);
   }
-};
-
-class QueryUser extends SfCommandletExecutor<{}> {
-  private username: string;
-  public constructor(username: string) {
-    super();
-    this.username = username;
-  }
-  public build(): Command {
-    return new SfCommandBuilder()
-      .withArg('data:query')
-      .withFlag('--query', `SELECT id FROM User WHERE username='${this.username}'`)
-      .withJson()
-      .withLogName('query_user')
-      .build();
-  }
-}
-
-class CreateDebugLevel extends SfCommandletExecutor<{}> {
-  public readonly developerName = `ReplayDebuggerLevels${Date.now()}`;
-  public build(): Command {
-    return new SfCommandBuilder()
-      .withArg('data:create:record')
-      .withFlag('--sobject', 'DebugLevel')
-      .withFlag(
-        '--values',
-        `developername=${this.developerName} MasterLabel=${this.developerName} apexcode=${APEX_CODE_DEBUG_LEVEL} visualforce=${VISUALFORCE_DEBUG_LEVEL}`
-      )
-      .withArg('--use-tooling-api')
-      .withJson()
-      .withLogName('create_debug_level')
-      .build();
-  }
-}
-
-class CreateTraceFlag extends SfCommandletExecutor<{}> {
-  private userId: string;
-
-  public constructor(userId: string) {
-    super();
-    this.userId = userId;
-  }
-
-  public build(): Command {
-    return new SfCommandBuilder()
-      .withArg('data:create:record')
-      .withFlag('--sobject', 'TraceFlag')
-      .withFlag(
-        '--values',
-        `tracedentityid='${
-          this.userId
-        }' logtype=developer_log debuglevelid=${developerLogTraceFlag.getDebugLevelId()} StartDate='' ExpirationDate='${developerLogTraceFlag
-          .getExpirationDate()
-          .toUTCString()}`
-      )
-      .withArg('--use-tooling-api')
-      .withJson()
-      .withLogName('create_trace_flag')
-      .build();
-  }
-}
-
-class UpdateDebugLevelsExecutor extends SfCommandletExecutor<{}> {
-  public build(): Command {
-    const nonNullDebugLevel = developerLogTraceFlag.getDebugLevelId()!;
-    return new SfCommandBuilder()
-      .withArg('data:update:record')
-      .withFlag('--sobject', 'DebugLevel')
-      .withFlag('--record-id', nonNullDebugLevel)
-      .withFlag('--values', `ApexCode=${APEX_CODE_DEBUG_LEVEL} Visualforce=${VISUALFORCE_DEBUG_LEVEL}`)
-      .withArg('--use-tooling-api')
-      .withJson()
-      .withLogName('update_debug_level')
-      .build();
-  }
-}
-
-class UpdateTraceFlagsExecutor extends SfCommandletExecutor<{}> {
-  public build(): Command {
-    const nonNullTraceFlag = developerLogTraceFlag.getTraceFlagId()!;
-    return new SfCommandBuilder()
-      .withArg('data:update:record')
-      .withFlag('--sobject', 'TraceFlag')
-      .withFlag('--record-id', nonNullTraceFlag)
-      .withFlag('--values', `StartDate='' ExpirationDate='${developerLogTraceFlag.getExpirationDate().toUTCString()}'`)
-      .withArg('--use-tooling-api')
-      .withJson()
-      .withLogName('update_trace_flag')
-      .build();
-  }
-}
-
-const workspaceChecker = new SfWorkspaceChecker();
-const parameterGatherer = new EmptyParametersGatherer();
-
-class QueryTraceFlag extends SfCommandletExecutor<{}> {
-  public build(userId: string): Command {
-    return new SfCommandBuilder()
-      .withDescription(nls.localize('start_apex_debug_logging'))
-      .withArg('data:query')
-      .withFlag(
-        '--query',
-        `SELECT id, logtype, startdate, expirationdate, debuglevelid, debuglevel.apexcode, debuglevel.visualforce FROM TraceFlag WHERE logtype='DEVELOPER_LOG' AND TracedEntityId='${userId}'`
-      )
-      .withArg('--use-tooling-api')
-      .withJson()
-      .withLogName('query_trace_flag')
-      .build();
-  }
-}
-
-export const startApexDebugLogging = async (): Promise<void> => {
-  const executor = new StartApexDebugLoggingExecutor();
-  const commandlet = new SfCommandlet(workspaceChecker, parameterGatherer, executor);
-  await commandlet.run();
 };
